% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/emod-package.R, R/emod.R
\docType{package}
\name{emod}
\alias{emod}
\alias{emod-package}
\title{emod}
\usage{
emod(loss, payroll, ...)
}
\arguments{
\item{loss}{S3 object of one of the loss classes defined in the \code{emod} package}

\item{payroll}{a data frame with
two columns.  The first column is the payroll class and the second column is the payroll
dollar value.  Payroll dollar value are for the entire 3 year period the NCCI emod takes
into account (i.e. do not segregate payroll dollars by policy year)}

\item{...}{additional arguments}
}
\description{
calculate NCCI style e-mod

generic function for calculating emods under various assumtions
}

