% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loss_ncci.R
\name{loss_ncci}
\alias{loss_ncci}
\title{loss_ncci}
\usage{
loss_ncci(year, type, incurred)
}
\arguments{
\item{year}{loss year}

\item{type}{loss type.  medical only "MO" or indemnity "IND"}

\item{incurred}{net incurred amount}
}
\description{
S3 class for insurer loss information necessary to calculate the
NCCI 3 year emod
}
\examples{
set.seed(1234)
loss_ncci(year = sample(c(2012, 2013, 2014), 20, replace = TRUE),
     type = sample(c("MO", "IND"), 20, replace = TRUE),
     incurred = rlnorm(20, 10, 2))
}

