% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loss_ncci.R
\name{summary.loss_ncci}
\alias{summary.loss_ncci}
\title{summary.loss_ncci}
\usage{
\method{summary}{loss_ncci}(loss_ncci)
}
\arguments{
\item{loss_ncci}{object of class loss_ncci}
}
\description{
Summarize loss_ncci by accident year
}
\examples{
test <- loss_ncci(year = sample(c(2011, 2012, 2013), 20, replace = TRUE),
                  type = sample(c("MO", "IND"), 20, replace = TRUE),
                  incurred = rlnorm(20, 10, 2))

set.seed(1234)
summary(test)
}

