% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/emod.R
\name{emod.loss_ncci}
\alias{emod.loss_ncci}
\title{NCCI emod}
\usage{
\method{emod}{loss_ncci}(loss, payroll, emod_year = 2015,
  primary_cut = 15500)
}
\description{
NCCI emod
}
\examples{
payroll <- list("2012" = data.frame(class = as.factor(c("8868", "9101")),
                                    payroll = c(10000000, 2000000)),
                "2013" = data.frame(class = as.factor(c("8868", "9101")),
                                    payroll = c(10500000, 2200000)),
                "2014" = data.frame(class = as.factor(c("8868", "9101")),
                                    payroll = c(11000000, 2400000))
            )

emod(test, payroll)
}

