% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/emod-package.R, R/emod.R
\docType{package}
\name{emod}
\alias{emod}
\alias{emod-package}
\title{emod}
\usage{
emod(loss, payroll, ...)
}
\arguments{
\item{loss}{S3 object of one of the loss classes defined in the \code{emod} package}

\item{payroll}{a list of n data frames.  Each list element contains a data frame with
two columns.  The first column is the payroll class and the second column is the
payroll dollar value.  Each data frame is in a seperate list item and contains a
years worth of payroll.}

\item{...}{additional arguments}
}
\description{
calculate NCCI style e-mod

generic function for calculating emods under various assumtions
}

